Change data

+++ app/src/main/java/com/weatherapplication/feature/cityweather/data/api/WeatherModelRemote.kt
@@ -8,7 +8,7 @@ class WeatherModelRemote(
 )
 
 data class Current(
-    @SerializedName("temp_c") val tempC: Int,
+    @SerializedName("temp_c") val tempC: Double,
     @SerializedName("temp_f") val tempF: Double,
     @SerializedName("is_day") val isDay: Int,
     @SerializedName("condition") val condition: Condition,
@@ -16,17 +16,17 @@ data class Current(
     @SerializedName("wind_kph") val windKph: Double,
     @SerializedName("wind_degree") val windDegree: Int,
     @SerializedName("wind_dir") val windDir: String,
-    @SerializedName("pressure_mb") val pressureMb: Int,
+    @SerializedName("pressure_mb") val pressureMb: Double,
     @SerializedName("pressure_in") val pressureIn: Double,
     @SerializedName("precip_mm") val precipMm: Double,
-    @SerializedName("precip_in") val precipIn: Int,
+    @SerializedName("precip_in") val precipIn: Double,
     @SerializedName("humidity") val humidity: Int,
     @SerializedName("cloud") val cloud: Int,
     @SerializedName("feelslike_c") val feelslikeC: Double,
-    @SerializedName("feelslike_f") val feelslikeF: Int,
-    @SerializedName("vis_km") val visKm: Int,
-    @SerializedName("vis_miles") val visMiles: Int,
-    @SerializedName("uv") val uv: Int,
+    @SerializedName("feelslike_f") val feelslikeF: Double,
+    @SerializedName("vis_km") val visKm: Double,
+    @SerializedName("vis_miles") val visMiles: Double,
+    @SerializedName("uv") val uv: Double,
     @SerializedName("gust_mph") val gustMph: Double,
     @SerializedName("gust_kph") val gustKph: Double
 )
@@ -60,11 +60,11 @@ data class Day(
     @SerializedName("maxwind_kph") val maxwindKph: Double,
     @SerializedName("totalprecip_mm") val totalprecipMm: Double,
     @SerializedName("totalprecip_in") val totalprecipIn: Double,
-    @SerializedName("avgvis_km") val avgvisKm: Int,
-    @SerializedName("avgvis_miles") val avgvisMiles: Int,
+    @SerializedName("avgvis_km") val avgvisKm: Double,
+    @SerializedName("avgvis_miles") val avgvisMiles: Double,
     @SerializedName("avghumidity") val avghumidity: Int,
     @SerializedName("condition") val condition: Condition,
-    @SerializedName("uv") val uv: Int
+    @SerializedName("uv") val uv: Double
 )
 
 data class Hour(
